refactor backdrop
